papsf.py
```python
# papsf.py

# Import necessary libraries
import os
import shutil
from jinja2 import Environment, FileSystemLoader
from git import Repo
import config

# Set up the Jinja2 environment
env = Environment(loader=FileSystemLoader(config.TEMPLATES_DIR))

def create_directory(path):
    """
    Create a directory at the given path.
    """
    os.makedirs(path, exist_ok=True)

def create_file(path, template_name, context={}):
    """
    Create a file at the given path using the specified template and context.
    """
    template = env.get_template(template_name)
    content = template.render(context)
    with open(path, 'w') as f:
        f.write(content)

def create_structure(base_dir, structure):
    """
    Create the project structure.
    """
    for key, value in structure.items():
        if isinstance(value, list):
            dir_path = os.path.join(base_dir, key)
            create_directory(dir_path)
            for item in value:
                if isinstance(item, str):
                    file_path = os.path.join(dir_path, item)
                    template_name = config.DEFAULT_FILE_TEMPLATES.get(item, 'default.j2')
                    create_file(file_path, template_name)
                elif isinstance(item, dict):
                    create_structure(dir_path, item)

def initialize_vcs(base_dir):
    """
    Initialize the version control system.
    """
    if config.VCS == 'git':
        Repo.init(base_dir)

def main():
    """
    Main function to start the Python Advanced Project Structure Fabricator.
    """
    # Create the base directory
    base_dir = os.path.join(config.OUTPUT_DIR, config.PROJECT_NAME)
    create_directory(base_dir)

    # Create the project structure
    create_structure(base_dir, config.DEFAULT_PROJECT_STRUCTURE)

    # Initialize the version control system
    initialize_vcs(base_dir)

    print(f"Project {config.PROJECT_NAME} has been successfully created at {base_dir}.")

if __name__ == "__main__":
    main()
```
