requirements.txt
# Required Python Libraries

# For general operations
numpy==1.21.2
pandas==1.3.3

# For AI-Driven Structural Optimization
tensorflow==2.6.0
keras==2.6.0
scikit-learn==0.24.2

# For Versatile Template Engine
jinja2==3.0.1

# For Seamless Version Control Integration
gitpython==3.1.24

# For Automated Documentation Production
sphinx==4.2.0
